Optimize WDIO configuration files
