+++ src/main/archetype/ui.tests/test-module/wdio.conf.local.js
@@ -14,83 +14,54 @@
  *  limitations under the License.
  */
 
-/*
- * WDIO Testrunner Configuration - See https://webdriver.io/docs/configurationfile.html
+/**
+ * DO NOT MODIFY
  */
-const conf = require('./lib/config');
-const commons = require('./lib/commons');
-const HtmlReporter = require('@rpii/wdio-html-reporter').HtmlReporter;
-const path = require('path');
-const log4js = require('log4js');
+let wdio_config = require('./wdio.conf.commons.js').config;
+let config = require('./lib/config');
 
-exports.config = {
-    runner: 'local',
+wdio_config.hostname = 'localhost';
+wdio_config.services = [
+    ['selenium-standalone', {
+        logPath: config.reports_path}
+    ]
+];
 
-    hostname: 'localhost',
-    // Tests
-    specs: [
-        './specs/**/*.js'
-    ],
+// Define capabilities based on configuration
+let capabilities = {};
 
-    services: [
-        ['selenium-standalone', {
-            logPath: conf.reports_path}
-        ]
-    ],
-
-    logLevel: 'debug',
-
-    bail: 0,
-
-    baseUrl: conf.aem.author.base_url,
-
-    sync: true,
-
-    waitforTimeout: 10000,
-    connectionRetryTimeout: 90000,
-    connectionRetryCount: 3,
-
-    framework: 'mocha',
-
-    // Location of the WDIO/Selenium logs
-    outputDir: conf.reports_path,
-
-    // Reporters
-    reporters: [
-        'spec',
-        ['junit', {
-            outputDir: path.join(conf.reports_path, 'junit'),
-            outputFileFormat: function(options) {
-                return `results-${options.cid}.${options.capabilities.browserName}.xml`;
-            }
-        }],
-        [HtmlReporter, {
-            debug: true,
-            outputDir: path.join(path.relative(process.cwd(), conf.reports_path), 'html/'),
-            filename: 'report.html',
-            reportTitle: 'UI Testing Basic Tests',
-            showInBrowser: false,
-            useOnAfterCommandForScreenshot: true,
-            LOG: log4js.getLogger('default')
-        }],
-    ],
-
-    // Mocha parameters
-    mochaOpts: {
-        ui: 'bdd',
-        timeout: 60000
-    },
-
-    // Gets executed before test execution begins
-    before: function() {
-        // Init custom WDIO commands (ex. AEMLogin)
-        require('./lib/wdio.commands');
-    },
-
-    // WDIO Hook executed after each test
-    afterTest: function() {
-        // Take a screenshot that will be attached in the HTML report
-        commons.takeScreenshot(browser);
+switch(config.selenium.browser) {
+case config.CHROME:
+    capabilities = {
+        maxInstances: 1,
+        browserName: 'chrome',
+        'goog:chromeOptions': {
+            'excludeSwitches': ['enable-automation']
+        }
+    };
+    if (config.selenium.headless === true) {
+        capabilities['goog:chromeOptions'].args = ['headless'];
     }
+    break;
+case config.FIREFOX:
+    capabilities = {
+        maxInstances: 1,
+        browserName: 'firefox',
+        'moz:firefoxOptions': {
+            prefs: {
+                // Prevent opening the extension tabs on startup
+                'extensions.enabledScopes': 0
+            }
+        }
+    };
+    if (config.selenium.headless === true) {
+        capabilities['moz:firefoxOptions'].args = ['-headless'];
+    }
+    break;
+default:
+    throw new Error('Unsupported browser defined in configuration!');
+}
 
-};
+wdio_config.capabilities = [capabilities];
+
+exports.config = wdio_config;
